fix: search has been fixed
Searched was 1 char behind because the search term was being added to state and then the term from state was being passed to the API. This bug was found in the review and fixed now, simply by just passing the term directly to the API.

Search results are being cleared if the added "letter" to the search term input field makes it so it doesn't match any books; if the results from the API search doesn't return any books, we catch the error and update state to an empty array.

Trim was removed to allow for more than one word to be searched. Ideally, trim would be implemented correctly, where it would remove empty spaces from the begging or end but not between words.